Move struct `header` and template struct `traceable` into `collector`
