+++ VERSION
@@ -1,1 +1,1 @@
-0.5.95
+0.5.96
